src/test-eventos/requirements.txt
pulsar-client
pulsar-client[avro]